requirements.txt
pandas==1.2.5
tqdm==4.61.2
imbalanced_learn==0.7.0
xgboost==1.3.3
numpy==1.20.2
imblearn==0.0
PyYAML==5.4.1
scikit_learn==0.24.2
